requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-braces==1.15.0
django-cleanup==6.0.0
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
sqlparse==0.4.2
whitenoise==6.2.0
